+++ rose_sdk/models/query.py
@@ -8,6 +8,7 @@ from .base import BaseModel
 
 class Query(BaseModel):
     """Query model."""
+
     pipeline_id: Optional[str] = None
     query_name: Optional[str] = None
     query_id: Optional[str] = None