+++ rose_sdk/models/role.py
@@ -8,6 +8,7 @@ from .base import BaseModel
 
 class Role(BaseModel):
     """Role model."""
+
     account_id: str
     role_name: str
     role_id: str
@@ -17,15 +18,18 @@ class Role(BaseModel):
 
 class RoleWithToken(Role):
     """Role model with access token."""
+
     token: str
 
 
 class CreateRoleRequest(BaseModel):
     """Request model for creating a role."""
+
     role_name: str
     permission: Optional[str] = None
 
 
 class UpdateRoleRequest(BaseModel):
     """Request model for updating a role."""
+
     permission: str